initial moving of code
