+++ Cargo.toml
@@ -4,4 +4,5 @@
 members = [
     "zkprotocol",
     "server",
+    "orgauth",
 ]
