+++ orgauth/Cargo.toml
@@ -1,26 +1,23 @@
 [package]
-name = "zknotes-server"
+name = "orgauth"
 version = "0.1.0"
-authors = ["Ben Burdette <bburdette@gmail.com>"]
-edition = "2018"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-zkprotocol = { path = "../zkprotocol" }
-orgauth = { path = "../orgauth" }
 base64 = "0.2.1"
 env_logger = "0.5.13"
 rusqlite = "0.23.0"
 log = "0.4.0"
 actix-web = "3.3.2"
-# actix-rt = "0.2.5"
 actix-files = "0.5.0"
 actix-session = "0.4.0"
-actix-cors = "0.5.4"
 futures = "0.1.25"
 serde = "1.0.124"
 serde_derive = "1.0.124"
 serde_json = "1.0.64"
-uuid = { version = "0.8", features = ["serde", "v4"] }
+# uuid = { version = "0.8", features = ["serde", "v4"] }
 crypto-hash = "0.3.1"
 time = "0.1"
 rand = "0.5.0"
