+++ server/src/main.rs
@@ -1,5 +1,5 @@
 mod config;
-mod email;
+// mod email;
 mod interfaces;
 mod search;
 mod sqldata;
@@ -119,106 +119,32 @@ fn user(
   }
 }
 
-fn register(data: web::Data<Config>, req: HttpRequest) -> HttpResponse {
-  info!("registration: uid: {:?}", req.match_info().get("uid"));
-  match sqldata::connection_open(data.db.as_path()) {
-    Ok(conn) => match (req.match_info().get("uid"), req.match_info().get("key")) {
-      (Some(uid), Some(key)) => {
-        // read user record.  does the reg key match?
-        match sqldata::read_user_by_name(&conn, uid) {
-          Ok(user) => {
-            if user.registration_key == Some(key.to_string()) {
-              let mut mu = user;
-              mu.registration_key = None;
-              match sqldata::update_user(&conn, &mu) {
-                Ok(_) => HttpResponse::Ok().body(
-                  format!(
-                    "<h1>You are registered!<h1> <a href=\"{}\">\
-                       Proceed to the main site</a>",
-                    data.mainsite
-                  )
-                  .to_string(),
-                ),
-                Err(_e) => HttpResponse::Ok().body("<h1>registration failed</h1>".to_string()),
-              }
-            } else {
-              HttpResponse::Ok().body("<h1>registration failed</h1>".to_string())
-            }
-          }
-          Err(_e) => HttpResponse::Ok().body("registration key or user doesn't match".to_string()),
+fn private(
+  session: Session,
+  data: web::Data<Config>,
+  item: web::Json<UserMessage>,
+  req: HttpRequest,
+) -> HttpResponse {
+  match session.get::<Uuid>("token")? {
+    None => Ok(ServerResponse {
+      what: "not logged in".to_string(),
+      content: serde_json::Value::Null,
+    }),
+    Some(token) => {
+      match sqldata::read_user_by_token(&conn, token, Some(config.login_token_expiration_ms)) {
+        Err(e) => {
+          info!("read_user_by_token error: {:?}", e);
+
+          Ok(ServerResponse {
+            what: "invalid user or pwd".to_string(),
+            content: serde_json::Value::Null,
+          })
+        }
+        Ok(userdata) => {
+          // finally!  processing messages as logged in user.
+          user_interface_loggedin(&config, userdata.id, &msg)
         }
       }
-      _ => HttpResponse::Ok().body("Uid, key not found!".to_string()),
-    },
-
-    Err(_e) => HttpResponse::Ok().body("<h1>registration failed</h1>".to_string()),
-  }
-}
-
-fn new_email(data: web::Data<Config>, req: HttpRequest) -> HttpResponse {
-  info!("new email: uid: {:?}", req.match_info().get("uid"));
-  match sqldata::connection_open(data.db.as_path()) {
-    Ok(conn) => match (req.match_info().get("uid"), req.match_info().get("token")) {
-      (Some(uid), Some(tokenstr)) => {
-        match Uuid::from_str(tokenstr) {
-          Err(_e) => HttpResponse::BadRequest().body("invalid token".to_string()),
-          Ok(token) => {
-            // read user record.  does the reg key match?
-            match sqldata::read_user_by_name(&conn, uid) {
-              Ok(user) => {
-                match sqldata::read_newemail(&conn, user.id, token) {
-                  Ok((email, tokendate)) => {
-                    match now() {
-                      Err(_e) => HttpResponse::InternalServerError()
-                        .body("<h1>'now' failed!</h1>".to_string()),
-
-                      Ok(now) => {
-                        if (now - tokendate) > data.email_token_expiration_ms {
-                          // TODO token expired?
-                          HttpResponse::UnprocessableEntity()
-                            .body("<h1>email change failed - token expired</h1>".to_string())
-                        } else {
-                          // put the email in the user record and update.
-                          let mut mu = user.clone();
-                          mu.email = email;
-                          match sqldata::update_user(&conn, &mu) {
-                            Ok(_) => {
-                              // delete the change email token record.
-                              match sqldata::remove_newemail(&conn, user.id, token) {
-                                Ok(_) => (),
-                                Err(e) => error!("error removing newemail record: {:?}", e),
-                              }
-                              HttpResponse::Ok().body(
-                                format!(
-                                  "<h1>Email address changed!<h1> <a href=\"{}\">\
-                                   Proceed to the main site</a>",
-                                  data.mainsite
-                                )
-                                .to_string(),
-                              )
-                            }
-                            Err(_e) => HttpResponse::InternalServerError()
-                              .body("<h1>email change failed</h1>".to_string()),
-                          }
-                        }
-                      }
-                    }
-                  }
-                  Err(_e) => HttpResponse::InternalServerError()
-                    .body("<h1>email change failed</h1>".to_string()),
-                }
-              }
-              Err(_e) => HttpResponse::BadRequest()
-                .body("email change token or user doesn't match".to_string()),
-            }
-          }
-        }
-      }
-      _ => HttpResponse::BadRequest().body("username or token not found!".to_string()),
-    },
-
-    Err(_e) => {
-      HttpResponse::InternalServerError().body("<h1>database connection failed</h1>".to_string())
     }
   }
 }
@@ -240,18 +166,6 @@ fn defcon() -> Config {
     reset_token_expiration_ms: 1 * 24 * 60 * 60 * 1000, // 1 day in milliseconds
     static_path: None,
   }
-}
-
-fn purge_tokens(config: &Config) -> Result<(), Box<dyn Error>> {
-  let conn = sqldata::connection_open(config.db.as_path())?;
-
-  sqldata::purge_login_tokens(&conn, config.login_token_expiration_ms)?;
-
-  sqldata::purge_email_tokens(&conn, config.email_token_expiration_ms)?;
-
-  sqldata::purge_reset_tokens(&conn, config.reset_token_expiration_ms)?;
-
-  Ok(())
 }
 
 fn load_config(filename: &str) -> Result<Config, Box<dyn Error>> {
@@ -388,6 +302,7 @@ async fn err_main() -> Result<(), Box<dyn Error>> {
                   .max_age(10 * 24 * 60 * 60), // 10 days
               )
               .service(web::resource("/public").route(web::post().to(public)))
+              .service(web::resource("/private").route(web::post().to(private)))
               .service(web::resource("/user").route(web::post().to(user)))
               .service(web::resource(r"/register/{uid}/{key}").route(web::get().to(register)))
               .service(web::resource(r"/newemail/{uid}/{token}").route(web::get().to(new_email)))
