+++ zkprotocol/src/content.rs
@@ -1,43 +1,6 @@
 #![allow(non_snake_case)]
 use uuid::Uuid;
 
-#[derive(Deserialize, Debug)]
-pub struct RegistrationData {
-  pub uid: String,
-  pub pwd: String,
-  pub email: String,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct Login {
-  pub uid: String,
-  pub pwd: String,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct ResetPassword {
-  pub uid: String,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct SetPassword {
-  pub uid: String,
-  pub newpwd: String,
-  pub reset_key: Uuid,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct ChangePassword {
-  pub oldpwd: String,
-  pub newpwd: String,
-}
-
-#[derive(Deserialize, Debug, Clone)]
-pub struct ChangeEmail {
-  pub pwd: String,
-  pub email: String,
-}
-
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct LoginData {
   pub userid: i64,
